pyproject.toml
[tool.poetry]
name = "fastapi-gym-crm"
version = "0.1.0"
description = ""
authors = ["Jimin Yoo <jmy0819_@naver.com>"]
readme = "README.md"
packages = [{ from = "src", include = "*" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
aiomysql = "^0.2.0"
greenlet = "^3.1.1"
cryptography = "^44.0.0"
pyjwt = "^2.10.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
pydantic = { version = "^2.10.4", extras = ["email"] }
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
